src/main/java/kr/megaptera/makaobank/services/TransactionService.java
package kr.megaptera.makaobank.services;

import kr.megaptera.makaobank.models.AccountNumber;
import kr.megaptera.makaobank.models.Transaction;
import kr.megaptera.makaobank.repositories.TransactionRepository;
import org.springframework.data.domain.PageRequest;
import org.springframework.data.domain.Pageable;
import org.springframework.data.domain.Sort;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@Service
@Transactional
public class TransactionService {
  private final TransactionRepository transactionRepository;

  public TransactionService(TransactionRepository transactionRepository) {
    this.transactionRepository = transactionRepository;
  }

  public List<Transaction> list(AccountNumber accountNumber, int page) {
    Sort sort = Sort.by("id").descending();
    Pageable pageable = PageRequest.of(page - 1, 100, sort);
    return transactionRepository
        .findAllBySenderOrReceiver(
            accountNumber, accountNumber, pageable);
  }
}
